refactor(routes): split simulation.py by resource (#235)
- Add _helpers.py with shared error_response
- Add runs.py: run list, run POST, run details, run turns
- Add agents.py: agents CRUD + agent follows CRUD
- Add metadata.py: feed-algorithms, metrics, config/default
- Add posts.py: posts list
- Reduce simulation.py to minimal composition module
- All endpoints remain at same paths; tests pass
