+++ simulation/api/routes/simulation.py
@@ -1,865 +1,20 @@
-"""Simulation run API routes."""
+"""Simulation API router composition.
 
-import asyncio
-import logging
+Aggregates resource-oriented route modules; keeps URL paths and response models stable.
+"""
 
-from fastapi import APIRouter, Depends, Query, Request
-from fastapi.responses import JSONResponse, Response
+from fastapi import APIRouter, Depends
 
-from lib.decorators import timed
-from lib.rate_limiting import RATE_LIMIT_AGENTS_CREATE, limiter
-from lib.request_logging import RunIdSource, log_route_completion_decorator
-from simulation.api.constants import (
-    DEFAULT_AGENT_LIST_LIMIT,
-    DEFAULT_AGENT_LIST_OFFSET,
-    DEFAULT_SIMULATION_CONFIG,
-    MAX_AGENT_LIST_LIMIT,
-)
 from simulation.api.dependencies.app_user import require_current_app_user
-from simulation.api.errors import (
-    ApiAgentFollowEdgeAlreadyExistsError,
-    ApiAgentFollowEdgeNotFoundError,
-    ApiAgentNotFoundError,
-    ApiHandleAlreadyExistsError,
-    ApiRunCreationFailedError,
-    ApiRunNotFoundError,
-    ApiSelfFollowNotAllowedError,
-    ApiTargetAgentNotFoundError,
-)
-from simulation.api.schemas.simulation import (
-    AgentFollowEdgeSchema,
-    AgentSchema,
-    CreateAgentFollowRequest,
-    CreateAgentRequest,
-    DefaultConfigSchema,
-    FeedAlgorithmSchema,
-    ListAgentFollowsResponse,
-    MetricSchema,
-    PostSchema,
-    RunDetailsResponse,
-    RunListItem,
-    RunRequest,
-    RunResponse,
-    TurnSchema,
-)
-from simulation.api.services.agent_command_service import create_agent, delete_agent
-from simulation.api.services.agent_follows_command_service import (
-    create_agent_follow,
-    delete_agent_follow,
-)
-from simulation.api.services.agent_follows_query_service import list_agent_follows
-from simulation.api.services.agent_query_service import list_agents
-from simulation.api.services.metadata_service import list_feed_algorithms, list_metrics
-from simulation.api.services.run_execution_service import execute
-from simulation.api.services.run_query_service import (
-    get_posts_by_ids,
-    get_run_details,
-    get_turns_for_run,
-    list_runs,
-)
-
-logger = logging.getLogger(__name__)
+from simulation.api.routes.agents import router as agents_router
+from simulation.api.routes.metadata import router as metadata_router
+from simulation.api.routes.posts import router as posts_router
+from simulation.api.routes.runs import router as runs_router
 
 router = APIRouter(
     tags=["simulation"], dependencies=[Depends(require_current_app_user)]
 )
-
-SIMULATION_RUN_ROUTE: str = "POST /v1/simulations/run"
-SIMULATION_RUNS_ROUTE: str = "GET /v1/simulations/runs"
-SIMULATION_RUN_DETAILS_ROUTE: str = "GET /v1/simulations/runs/{run_id}"
-SIMULATION_RUN_TURNS_ROUTE: str = "GET /v1/simulations/runs/{run_id}/turns"
-SIMULATION_AGENTS_ROUTE: str = "GET /v1/simulations/agents"
-SIMULATION_AGENTS_CREATE_ROUTE: str = "POST /v1/simulations/agents"
-SIMULATION_AGENTS_DELETE_ROUTE: str = "DELETE /v1/simulations/agents/{handle}"
-SIMULATION_AGENT_FOLLOWS_ROUTE: str = "GET /v1/simulations/agents/{handle}/follows"
-SIMULATION_AGENT_FOLLOWS_CREATE_ROUTE: str = (
-    "POST /v1/simulations/agents/{handle}/follows"
-)
-SIMULATION_AGENT_FOLLOWS_DELETE_ROUTE: str = (
-    "DELETE /v1/simulations/agents/{handle}/follows/{target_handle}"
-)
-SIMULATION_POSTS_ROUTE: str = "GET /v1/simulations/posts"
-SIMULATION_FEED_ALGORITHMS_ROUTE: str = "GET /v1/simulations/feed-algorithms"
-SIMULATION_METRICS_ROUTE: str = "GET /v1/simulations/metrics"
-SIMULATION_CONFIG_DEFAULT_ROUTE: str = "GET /v1/simulations/config/default"
-
-
-@router.get(
-    "/simulations/feed-algorithms",
-    response_model=list[FeedAlgorithmSchema],
-    status_code=200,
-    summary="List feed algorithms",
-    description="Return available feed algorithms with metadata for the UI.",
-)
-@log_route_completion_decorator(
-    route=SIMULATION_FEED_ALGORITHMS_ROUTE, success_type=list
-)
-async def get_simulation_feed_algorithms(
-    request: Request,
-) -> list[FeedAlgorithmSchema] | Response:
-    """Return registered feed algorithms with metadata."""
-    return await _execute_get_feed_algorithms(request)
-
-
-@router.get(
-    "/simulations/metrics",
-    response_model=list[MetricSchema],
-    status_code=200,
-    summary="List metrics",
-    description="Return available metrics with metadata for the UI.",
-)
-@log_route_completion_decorator(route=SIMULATION_METRICS_ROUTE, success_type=list)
-async def get_simulation_metrics(
-    request: Request,
-) -> list[MetricSchema] | Response:
-    """Return registered metrics with metadata."""
-    return await _execute_get_metrics(request)
-
-
-@router.get(
-    "/simulations/config/default",
-    response_model=DefaultConfigSchema,
-    status_code=200,
-    summary="Get default simulation config",
-    description="Return default config for simulation start form (num_agents, num_turns, metric_keys).",
-)
-@log_route_completion_decorator(
-    route=SIMULATION_CONFIG_DEFAULT_ROUTE, success_type=DefaultConfigSchema
-)
-async def get_simulation_config_default(
-    request: Request,
-) -> DefaultConfigSchema | Response:
-    """Return default config for simulation start form."""
-    return await _execute_get_default_config(request)
-
-
-@router.post(
-    "/simulations/agents",
-    response_model=AgentSchema,
-    status_code=201,
-    summary="Create simulation agent",
-    description="Create a user-generated agent.",
-)
-@limiter.limit(RATE_LIMIT_AGENTS_CREATE)
-@log_route_completion_decorator(
-    route=SIMULATION_AGENTS_CREATE_ROUTE, success_type=AgentSchema
-)
-async def post_simulation_agents(
-    request: Request, body: CreateAgentRequest
-) -> AgentSchema | Response:
-    """Create an agent and return it."""
-    return await _execute_post_simulation_agents(request, body=body)
-
-
-@router.get(
-    "/simulations/agents",
-    response_model=list[AgentSchema],
-    status_code=200,
-    summary="List simulation agents",
-    description="Return simulation agent profiles from DB for View agents and Create form.",
-)
-@log_route_completion_decorator(route=SIMULATION_AGENTS_ROUTE, success_type=list)
-async def get_simulation_agents(
-    request: Request,
-    q: str | None = Query(
-        default=None,
-        max_length=200,
-        description=(
-            "Optional handle search query (case-insensitive substring). "
-            "Supports '*' (any-length) and '?' (single-character) wildcards."
-        ),
-    ),
-    limit: int = Query(
-        default=DEFAULT_AGENT_LIST_LIMIT,
-        ge=1,
-        le=MAX_AGENT_LIST_LIMIT,
-        description=(
-            "Maximum number of agents to return (ordered by updated_at DESC, handle ASC)."
-        ),
-    ),
-    offset: int = Query(
-        default=DEFAULT_AGENT_LIST_OFFSET,
-        ge=0,
-        description=(
-            "Number of agents to skip before returning results (ordered by updated_at DESC, handle ASC)."
-        ),
-    ),
-) -> list[AgentSchema] | Response:
-    """Return all simulation agents from the database."""
-    return await _execute_get_simulation_agents(
-        request, q=q, limit=limit, offset=offset
-    )
-
-
-@router.get(
-    "/simulations/agents/{handle}/follows",
-    response_model=ListAgentFollowsResponse,
-    status_code=200,
-    summary="List seed-state follows for an agent",
-    description="Return editable pre-run follow edges for the given agent handle.",
-)
-@log_route_completion_decorator(
-    route=SIMULATION_AGENT_FOLLOWS_ROUTE,
-    success_type=ListAgentFollowsResponse,
-)
-async def get_simulation_agent_follows(
-    request: Request,
-    handle: str,
-    limit: int = Query(
-        default=DEFAULT_AGENT_LIST_LIMIT,
-        ge=1,
-        le=MAX_AGENT_LIST_LIMIT,
-        description="Maximum number of follow edges to return.",
-    ),
-    offset: int = Query(
-        default=DEFAULT_AGENT_LIST_OFFSET,
-        ge=0,
-        description="Number of follow edges to skip before returning results.",
-    ),
-) -> ListAgentFollowsResponse | Response:
-    """Return paginated editable follow edges for the given agent handle."""
-    return await _execute_get_simulation_agent_follows(
-        request,
-        handle=handle,
-        limit=limit,
-        offset=offset,
-    )
-
-
-@router.post(
-    "/simulations/agents/{handle}/follows",
-    response_model=AgentFollowEdgeSchema,
-    status_code=201,
-    summary="Create seed-state follow edge",
-    description="Create an editable pre-run follow edge between two internal agents.",
-)
-@log_route_completion_decorator(
-    route=SIMULATION_AGENT_FOLLOWS_CREATE_ROUTE,
-    success_type=AgentFollowEdgeSchema,
-)
-async def post_simulation_agent_follows(
-    request: Request,
-    handle: str,
-    body: CreateAgentFollowRequest,
-) -> AgentFollowEdgeSchema | Response:
-    """Create one editable follow edge for the given agent handle."""
-    return await _execute_post_simulation_agent_follows(
-        request,
-        handle=handle,
-        body=body,
-    )
-
-
-@router.delete(
-    "/simulations/agents/{handle}/follows/{target_handle}",
-    status_code=204,
-    summary="Delete seed-state follow edge",
-    description="Delete an editable pre-run follow edge between two internal agents.",
-)
-@log_route_completion_decorator(
-    route=SIMULATION_AGENT_FOLLOWS_DELETE_ROUTE,
-    success_type=type(None),
-)
-async def delete_simulation_agent_follow(
-    request: Request,
-    handle: str,
-    target_handle: str,
-) -> Response:
-    """Delete one editable follow edge for the given agent handle."""
-    return await _execute_delete_simulation_agent_follow(
-        request,
-        handle=handle,
-        target_handle=target_handle,
-    )
-
-
-@router.delete(
-    "/simulations/agents/{handle}",
-    status_code=204,
-    summary="Delete simulation agent",
-    description="Delete a simulation agent by handle.",
-)
-@log_route_completion_decorator(
-    route=SIMULATION_AGENTS_DELETE_ROUTE, success_type=type(None)
-)
-async def delete_simulation_agent(request: Request, handle: str) -> Response:
-    """Delete an agent and its related data."""
-    return await _execute_delete_simulation_agent(request, handle=handle)
-
-
-@router.get(
-    "/simulations/runs",
-    response_model=list[RunListItem],
-    status_code=200,
-    summary="List simulation runs",
-    description="Return simulation run summaries for the UI.",
-)
-@log_route_completion_decorator(route=SIMULATION_RUNS_ROUTE, success_type=list)
-async def get_simulation_runs(request: Request) -> list[RunListItem] | Response:
-    """Return all simulation runs from the database (app engine backed by SqliteTransactionProvider; DB path from SIM_DB_PATH or local dev DB in LOCAL mode)."""
-    return await _execute_get_simulation_runs(request)
-
-
-@router.post(
-    "/simulations/run",
-    response_model=RunResponse,
-    status_code=200,
-    summary="Run a simulation",
-    description="Execute a synchronous simulation run.",
-)
-@limiter.limit("5/minute")
-@log_route_completion_decorator(
-    route=SIMULATION_RUN_ROUTE,
-    success_type=RunResponse,
-    run_id_from=RunIdSource.RESPONSE,
-)
-async def post_simulations_run(
-    request: Request, body: RunRequest
-) -> RunResponse | Response:
-    """Execute a simulation run and return completed or partial results."""
-    return await _execute_simulation_run(request=request, body=body)
-
-
-@router.get(
-    "/simulations/runs/{run_id}",
-    response_model=RunDetailsResponse,
-    status_code=200,
-    summary="Get simulation run details",
-    description="Fetch run config and turn-by-turn action summary by run ID.",
-)
-@log_route_completion_decorator(
-    route=SIMULATION_RUN_DETAILS_ROUTE,
-    success_type=RunDetailsResponse,
-    run_id_from=RunIdSource.RESPONSE,
-)
-async def get_simulation_run(
-    request: Request, run_id: str
-) -> RunDetailsResponse | Response:
-    """Return run details and turn history for a persisted run."""
-    return await _execute_get_simulation_run(request=request, run_id=run_id)
-
-
-@router.get(
-    "/simulations/posts",
-    response_model=list[PostSchema],
-    status_code=200,
-    summary="List simulation posts",
-    description="Return posts, optionally filtered by canonical post_ids. Batch lookup for feed resolution.",
-)
-@log_route_completion_decorator(route=SIMULATION_POSTS_ROUTE, success_type=list)
-async def get_simulation_posts(
-    request: Request,
-    post_ids: list[str] | None = Query(
-        default=None, description="Filter by canonical post_ids"
-    ),
-) -> list[PostSchema] | Response:
-    """Return posts from the database (via SqliteTransactionProvider; DB path from SIM_DB_PATH or local dev DB in LOCAL mode)."""
-    return await _execute_get_simulation_posts(request, post_ids=post_ids)
-
-
-@router.get(
-    "/simulations/runs/{run_id}/turns",
-    response_model=dict[str, TurnSchema],
-    status_code=200,
-    summary="Get simulation run turns",
-    description="Return full per-turn payload for a run ID.",
-)
-@log_route_completion_decorator(
-    route=SIMULATION_RUN_TURNS_ROUTE, success_type=dict, run_id_from=RunIdSource.PATH
-)
-async def get_simulation_run_turns(
-    request: Request, run_id: str
-) -> dict[str, TurnSchema] | Response:
-    """Return turn payload for a run from the database (via SqliteTransactionProvider; DB path from SIM_DB_PATH or local dev DB in LOCAL mode)."""
-    return await _execute_get_simulation_run_turns(request, run_id=run_id)
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_get_metrics(
-    request: Request,
-) -> list[MetricSchema] | Response:
-    """Fetch metrics and convert unexpected failures to HTTP responses."""
-    try:
-        return await asyncio.to_thread(list_metrics)
-    except Exception:
-        logger.exception("Unexpected error while listing metrics")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_get_feed_algorithms(
-    request: Request,
-) -> list[FeedAlgorithmSchema] | Response:
-    """Fetch feed algorithms and convert unexpected failures to HTTP responses."""
-    try:
-        return await asyncio.to_thread(list_feed_algorithms)
-    except Exception:
-        logger.exception("Unexpected error while listing feed algorithms")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_get_default_config(
-    request: Request,
-) -> DefaultConfigSchema | Response:
-    """Fetch default config and convert unexpected failures to HTTP responses."""
-    try:
-        return DEFAULT_SIMULATION_CONFIG
-    except Exception:
-        logger.exception("Unexpected error while fetching default config")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_get_simulation_posts(
-    request: Request,
-    *,
-    post_ids: list[str] | None = None,
-) -> list[PostSchema] | Response:
-    """Fetch posts and convert unexpected failures to HTTP responses."""
-    try:
-        engine = request.app.state.engine
-        return await asyncio.to_thread(
-            get_posts_by_ids, post_ids=post_ids, engine=engine
-        )
-    except Exception:
-        logger.exception("Unexpected error while listing simulation posts")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_get_simulation_runs(
-    request: Request,
-) -> list[RunListItem] | Response:
-    """Fetch run summaries from the database and convert failures to HTTP responses."""
-    try:
-        engine = request.app.state.engine
-        # Use to_thread for consistency with other async routes and to prepare for real I/O later.
-        return await asyncio.to_thread(list_runs, engine=engine)
-    except Exception:
-        logger.exception("Unexpected error while listing simulation runs")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_post_simulation_agents(
-    request: Request, *, body: CreateAgentRequest
-) -> AgentSchema | Response:
-    """Create agent and convert known failures to HTTP responses."""
-    try:
-        app_state = request.app.state
-        return await asyncio.to_thread(
-            create_agent,
-            body,
-            transaction_provider=app_state.transaction_provider,
-            agent_repo=app_state.agent_repo,
-            bio_repo=app_state.agent_bio_repo,
-            metadata_repo=app_state.agent_metadata_repo,
-        )
-    except ApiHandleAlreadyExistsError as e:
-        return _error_response(
-            status_code=409,
-            code="HANDLE_ALREADY_EXISTS",
-            message="Agent with this handle already exists",
-            detail=e.handle,
-        )
-    except ValueError as e:
-        return _error_response(
-            status_code=422,
-            code="VALIDATION_ERROR",
-            message=str(e),
-            detail=None,
-        )
-    except Exception:
-        logger.exception("Unexpected error while creating agent")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_get_simulation_agents(
-    request: Request,
-    *,
-    q: str | None,
-    limit: int,
-    offset: int,
-) -> list[AgentSchema] | Response:
-    """Fetch agent list from DB and convert unexpected failures to HTTP responses."""
-    try:
-        app_state = request.app.state
-        return await asyncio.to_thread(
-            list_agents,
-            agent_repo=app_state.agent_repo,
-            bio_repo=app_state.agent_bio_repo,
-            metadata_repo=app_state.agent_metadata_repo,
-            q=q,
-            limit=limit,
-            offset=offset,
-        )
-    except Exception:
-        logger.exception("Unexpected error while listing simulation agents")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_get_simulation_agent_follows(
-    request: Request,
-    *,
-    handle: str,
-    limit: int,
-    offset: int,
-) -> ListAgentFollowsResponse | Response:
-    """Fetch seed-state follow edges and convert known failures to HTTP responses."""
-    try:
-        app_state = request.app.state
-        return await asyncio.to_thread(
-            list_agent_follows,
-            handle,
-            agent_repo=app_state.agent_repo,
-            agent_follow_edge_repo=app_state.agent_follow_edge_repo,
-            limit=limit,
-            offset=offset,
-        )
-    except ApiAgentNotFoundError as e:
-        return _error_response(
-            status_code=404,
-            code="AGENT_NOT_FOUND",
-            message="Agent not found",
-            detail=e.handle,
-        )
-    except ValueError as e:
-        return _error_response(
-            status_code=422,
-            code="VALIDATION_ERROR",
-            message=str(e),
-            detail=None,
-        )
-    except Exception:
-        logger.exception("Unexpected error while listing simulation agent follows")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_post_simulation_agent_follows(
-    request: Request,
-    *,
-    handle: str,
-    body: CreateAgentFollowRequest,
-) -> AgentFollowEdgeSchema | Response:
-    """Create seed-state follow edge and convert known failures to HTTP responses."""
-    try:
-        app_state = request.app.state
-        return await asyncio.to_thread(
-            create_agent_follow,
-            handle,
-            body,
-            transaction_provider=app_state.transaction_provider,
-            agent_repo=app_state.agent_repo,
-            agent_follow_edge_repo=app_state.agent_follow_edge_repo,
-            metadata_repo=app_state.agent_metadata_repo,
-        )
-    except ApiAgentNotFoundError as e:
-        return _error_response(
-            status_code=404,
-            code="AGENT_NOT_FOUND",
-            message="Agent not found",
-            detail=e.handle,
-        )
-    except ApiTargetAgentNotFoundError as e:
-        return _error_response(
-            status_code=404,
-            code="TARGET_AGENT_NOT_FOUND",
-            message="Target agent not found",
-            detail=e.handle,
-        )
-    except ApiAgentFollowEdgeAlreadyExistsError as e:
-        return _error_response(
-            status_code=409,
-            code="FOLLOW_EDGE_ALREADY_EXISTS",
-            message="Follow edge already exists",
-            detail=f"{e.follower_handle}->{e.target_handle}",
-        )
-    except ApiSelfFollowNotAllowedError as e:
-        return _error_response(
-            status_code=422,
-            code="SELF_FOLLOW_NOT_ALLOWED",
-            message="Agent cannot follow itself",
-            detail=e.handle,
-        )
-    except ValueError as e:
-        return _error_response(
-            status_code=422,
-            code="VALIDATION_ERROR",
-            message=str(e),
-            detail=None,
-        )
-    except Exception:
-        logger.exception("Unexpected error while creating simulation agent follow edge")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_delete_simulation_agent_follow(
-    request: Request,
-    *,
-    handle: str,
-    target_handle: str,
-) -> Response:
-    """Delete seed-state follow edge and convert known failures to HTTP responses."""
-    try:
-        app_state = request.app.state
-        await asyncio.to_thread(
-            delete_agent_follow,
-            handle,
-            target_handle,
-            transaction_provider=app_state.transaction_provider,
-            agent_repo=app_state.agent_repo,
-            agent_follow_edge_repo=app_state.agent_follow_edge_repo,
-            metadata_repo=app_state.agent_metadata_repo,
-        )
-        return Response(status_code=204)
-    except ApiAgentNotFoundError as e:
-        return _error_response(
-            status_code=404,
-            code="AGENT_NOT_FOUND",
-            message="Agent not found",
-            detail=e.handle,
-        )
-    except ApiTargetAgentNotFoundError as e:
-        return _error_response(
-            status_code=404,
-            code="TARGET_AGENT_NOT_FOUND",
-            message="Target agent not found",
-            detail=e.handle,
-        )
-    except ApiAgentFollowEdgeNotFoundError as e:
-        return _error_response(
-            status_code=404,
-            code="FOLLOW_EDGE_NOT_FOUND",
-            message="Follow edge not found",
-            detail=f"{e.follower_handle}->{e.target_handle}",
-        )
-    except ValueError as e:
-        return _error_response(
-            status_code=422,
-            code="VALIDATION_ERROR",
-            message=str(e),
-            detail=None,
-        )
-    except Exception:
-        logger.exception("Unexpected error while deleting simulation agent follow edge")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_delete_simulation_agent(
-    request: Request, *, handle: str
-) -> Response:
-    """Delete agent and convert known failures to HTTP responses."""
-    try:
-        app_state = request.app.state
-        await asyncio.to_thread(
-            delete_agent,
-            handle,
-            transaction_provider=app_state.transaction_provider,
-            agent_repo=app_state.agent_repo,
-            bio_repo=app_state.agent_bio_repo,
-            agent_follow_edge_repo=app_state.agent_follow_edge_repo,
-            metadata_repo=app_state.agent_metadata_repo,
-        )
-        return Response(status_code=204)
-    except ApiAgentNotFoundError as e:
-        return _error_response(
-            status_code=404,
-            code="AGENT_NOT_FOUND",
-            message="Agent not found",
-            detail=e.handle,
-        )
-    except ValueError as e:
-        return _error_response(
-            status_code=422,
-            code="VALIDATION_ERROR",
-            message=str(e),
-            detail=None,
-        )
-    except Exception:
-        logger.exception("Unexpected error while deleting agent")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_simulation_run(
-    request: Request, body: RunRequest
-) -> RunResponse | Response:
-    """Run the simulation and return response; used for timing and logging."""
-    engine = request.app.state.engine
-    current_app_user = getattr(request.state, "current_app_user")
-    if current_app_user is None:
-        raise RuntimeError(
-            "current_app_user was not set on request.state, but is required for simulation run creation."
-        )
-    created_by_app_user_id = current_app_user.id
-    try:
-        return await asyncio.to_thread(
-            execute,
-            request=body,
-            engine=engine,
-            created_by_app_user_id=created_by_app_user_id,
-        )
-    except ApiRunCreationFailedError as e:
-        logger.exception("Simulation run failed before run creation")
-        return _error_response(
-            status_code=500,
-            code="RUN_CREATION_FAILED",
-            message=e.message,
-            detail=None,
-        )
-    except Exception:
-        logger.exception("Unexpected error during simulation run")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_get_simulation_run_turns(
-    request: Request,
-    run_id: str,
-) -> dict[str, TurnSchema] | Response:
-    """Fetch run turns and convert known failures to HTTP responses."""
-    try:
-        engine = request.app.state.engine
-        return await asyncio.to_thread(get_turns_for_run, run_id=run_id, engine=engine)
-    except ApiRunNotFoundError as e:
-        return _error_response(
-            status_code=404,
-            code="RUN_NOT_FOUND",
-            message="Run not found",
-            detail=e.run_id,
-        )
-    except ValueError as e:
-        return _error_response(
-            status_code=400,
-            code="INVALID_RUN_ID",
-            message="Invalid run_id",
-            detail=str(e),
-        )
-    except Exception:
-        logger.exception("Unexpected error while fetching run turns")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-@timed(attach_attr="duration_ms", log_level=None)
-async def _execute_get_simulation_run(
-    request: Request, run_id: str
-) -> RunDetailsResponse | Response:
-    """Fetch persisted run details and convert known failures to HTTP responses."""
-    engine = request.app.state.engine
-    try:
-        return await asyncio.to_thread(
-            get_run_details,
-            run_id=run_id,
-            engine=engine,
-        )
-    except ApiRunNotFoundError as e:
-        return _error_response(
-            status_code=404,
-            code="RUN_NOT_FOUND",
-            message="Run not found",
-            detail=e.run_id,
-        )
-    except ValueError as e:
-        return _error_response(
-            status_code=400,
-            code="INVALID_RUN_ID",
-            message="Invalid run_id",
-            detail=str(e),
-        )
-    except Exception:
-        logger.exception("Unexpected error while fetching run details")
-        return _error_response(
-            status_code=500,
-            code="INTERNAL_ERROR",
-            message="Internal server error",
-            detail=None,
-        )
-
-
-def _error_response(
-    status_code: int,
-    code: str,
-    message: str,
-    detail: str | None = None,
-) -> JSONResponse:
-    return JSONResponse(
-        status_code=status_code,
-        content={
-            "error": {
-                "code": code,
-                "message": message,
-                "detail": detail,
-            }
-        },
-    )
+router.include_router(runs_router)
+router.include_router(agents_router)
+router.include_router(metadata_router)
+router.include_router(posts_router)
